Changes
Whew!

Okay, so here’s what I did:

I added an addBook method to the User’s class so that books could be
added to a user’s getOurBooks variable.

Also. I had to change getOurBooks so that it basically loaded from the
MasterList textfile and decides whether or nor each book is a specified
user’s book.

Also a result of that last edit, I had to throw a bunch of IOExceptions
to the BookList class.